# Conflicts:
#	include/ptr_/SharedObjPtr.h

+++ test/example_simple.cpp
@@ -21,7 +21,7 @@ int main() {
     // Set values in the inner tag.
     tag->setString("a_string",std::string("some string"));
     double* doubarr = new double[20];
-    for (int i=0; i<20; ++i) {
+    for (size_t i=0; i<20; ++i) {
         doubarr[i] = i*0.13413723472374e-15;
         std::cout << doubarr[i] << ' ';
     }
